그룹 함수 적용 하여 리팩토링 -> BookService 통계관련 기능

+++ src/main/kotlin/com/group/libraryapp/controller/book/BookController.kt
@@ -32,7 +32,7 @@ class BookController(
     }
 
     @GetMapping("/book/loan")
-    fun countLoanedBool(): Int {
+    fun countLoanedBool(): Long {
         return bookService.countLoanedBook();
     }
